chatbot/dataset/dataflow/requirements.txt
kazoo
pyquery
html2text
thrift
kafka-python
retrying
requests
ujson
pyformance
